Introduce RDGOptionalDatastructures
RDGOptionalDatastructures are datastructures that may be appended to an RDG
These datastructures are represented by a manifest file, which is json
encoded and stores the information required to load the optional
datastructure.

The only mandatory field in an optional datastructure manifest file is
"paths" which is used to list any additional files that make up the
optional datastructure. The "paths" field is
```
unordered_map<name, file_path>
```
the file_path is relative to the rdg_directory.

The RDGPartHeader tracks these optional data structures in the new field
```
optional_datastructure_manifests_
```
Any amount of optional datastructures may or may not be present in any
RDG. It is expected that if an optional datastructure is not present,
that the user will generate it instead.

The first optional datastructures are RDKLSHIndexPrimitive and
RDKSubstructureIndexPrimitive.

JIRA: KAT-3696